Basic addition with floating point numbers

+++ Code/2_Calculator/calculator.py
@@ -1,3 +1,3 @@
-x = int(input('What is x? '))
-y = int(input('What is y? '))
+x = float(input('What is x? '))
+y = float(input('What is y? '))
 print(f'{x} plus {y} = {x + y}')
